backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pydantic
python-multipart
Jinja2
